Unicode bugs fixed, added DELETE_CSV_AFTER_SPLIT

+++ split_csv.py
@@ -9,15 +9,16 @@ SOURCE_FILE = settings.csv_options['file']
 CHUNK_FILE_PREFIX = settings.csv_options['chunck_prefix']
 CHUNK_DEST_DIR = settings.csv_options['chunk_dir']
 LINES_PER_CHUNK = settings.csv_options['lines_per_chunk']
-
+DELETE_CSV_AFTER_SPLIT = settings.delete_csv_after_split
 
 def main(source_file, dest_dir, chunk_prefix, lines_per_chunk):
     try:
         split_csv(source_file, dest_dir, chunk_prefix, lines_per_chunk)
-    except:
-        sys.exit('Error splitting CSV file.')
-    print(f'Deleting original file: {source_file}...')
-    os.remove(source_file)
+    except Exception as error:
+        sys.exit(f'Error splitting CSV file: \n\n{error}')
+    if DELETE_CSV_AFTER_SPLIT:
+        print(f'Deleting original file: {source_file}...')
+        os.remove(source_file)
 
 
 def split_csv(source_file, dest_dir, chunk_prefix, lines_per_chunk):
@@ -25,11 +26,12 @@ def split_csv(source_file, dest_dir, chunk_prefix, lines_per_chunk):
     destination directory.
     """
     start_num = get_start_number(dest_dir, chunk_prefix)
-    with open(source_file, 'rt') as f:
+    with open(source_file, 'rt', encoding='utf8') as f:
         for file_num, chunk in enumerate(
             read_chunck(f, lines_per_chunk), start=start_num):
             dest_file = dest_dir / f'{chunk_prefix}{file_num:03d}.csv'
-            dest_file.write_text(chunk)
+            print(f'Writing to file {dest_file}...')
+            dest_file.write_text(chunk, encoding='utf8')
 
 
 def read_chunck(file, number_of_lines):
@@ -42,27 +44,7 @@ def read_chunck(file, number_of_lines):
             chunk = []
     else:
         yield ''.join(chunk)  
-        
-    # while True:
-    #     chunk = file.readlines(number_of_lines)
-    #     if not chunk:
-    #         break
-    #     else:
-    #         yield chunk
-    # lines = []
-    # i = 0
-    # while True:
-    #     try:
-    #         lines.append(file.readline())
-    #         i += 1
-    #     except:
-    #         yield ''.join(lines)
-    #         break
-    #     if i >= number_of_lines:
-    #         yield ''.join(lines)
-    #         i = 0
-    #         lines = []
-            
+
 
 def get_start_number(dest_dir, chunk_prefix):
     """Return the first free chunk number. Existing chunks are left
